docs: update storybook doc status, switchbutton and text
