+++ src/components/Status/__stories__/Playground.stories.tsx
@@ -1,0 +1,6 @@
+import { Template } from './Template.stories'
+
+export const Playground = Template.bind({})
+Playground.args = {
+  variant: 'success',
+}
