+++ src/components/Status/__stories__/Template.stories.tsx
@@ -1,0 +1,7 @@
+import { Story } from '@storybook/react'
+import { ComponentProps } from 'react'
+import Status from '../index'
+
+export const Template: Story<ComponentProps<typeof Status>> = args => (
+  <Status {...args} />
+)
