+++ src/components/SwitchButton/__stories__/Playground.stories.tsx
@@ -1,0 +1,14 @@
+import { Template } from './Template.stories'
+
+export const Playground = Template.bind({})
+Playground.args = {
+  leftButton: {
+    label: 'Left',
+    value: 'left',
+  },
+  rightButton: {
+    label: 'Right',
+    value: 'right',
+  },
+  value: 'left',
+}
