+++ src/components/Text/__stories__/Disabled.stories.tsx
@@ -1,0 +1,9 @@
+import { Template } from './Template.stories'
+
+export const Disabled = Template.bind({})
+Disabled.args = {
+  as: 'div',
+  children: 'This is a basic Text',
+  disabled: true,
+  variant: 'body',
+}
