+++ src/components/Text/__stories__/Playground.stories.tsx
@@ -1,0 +1,8 @@
+import { Template } from './Template.stories'
+
+export const Playground = Template.bind({})
+Playground.args = {
+  as: 'div',
+  children: 'This is a basic Text',
+  variant: 'body',
+}
