+++ src/components/Text/__stories__/Template.stories.tsx
@@ -1,0 +1,7 @@
+import { Story } from '@storybook/react'
+import { ComponentProps } from 'react'
+import Text from '../index'
+
+export const Template: Story<ComponentProps<typeof Text>> = args => (
+  <Text {...args} />
+)
