+++ src/components/Text/__stories__/Variants.stories.tsx
@@ -1,0 +1,21 @@
+import { ComponentStory } from '@storybook/react'
+import Text, { textVariants } from '../index'
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export const Variants: ComponentStory<typeof Text> = props => (
+  <>
+    {textVariants.map(variant => (
+      <Text as="div" key={variant} variant={variant}>
+        {variant}
+      </Text>
+    ))}
+  </>
+)
+
+Variants.parameters = {
+  docs: {
+    description: {
+      story: 'Set variant using `variant` property.',
+    },
+  },
+}
